pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "confidantic"
version = "0.2.0"
description = "Your most trusted confidant for app & library settings."
readme = "README.md"
authors = [{name = "Bullish-Design", email = "BullishDesignEngineering@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3",
    "rich>=13.6.0",
]

[project.optional-dependencies]
git = ["GitPython>=3.1.43"]

[project.urls]
Repository = "https://github.com/Bullish-Design/confidantic"

[tool.hatch.build.targets.wheel]
packages = ["src/confidantic"]

[project.scripts]
conf = "confidantic.cli:main"

